@@ -1,3 +1,4 @@
+import 'reflect-metadata';
 import { injectable, inject } from 'inversify';
 import { Constants } from 'eris';
 import { Module } from '../../util';
@@ -6,10 +7,12 @@ import findUser from '../../util/UserUtil';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
 import Bot from '../../structures/Bot';
+import WarningService from '../../structures/services/WarningService';
 
 @injectable()
 export default class WarningsCommand extends Command {
-  constructor(@inject(TYPES.Bot) client: Bot) {
+  constructor(@inject(TYPES.Bot) client: Bot,
+              @inject(TYPES.WarningService) private warningService: WarningService) {
     super(client, {
       name: 'warnings',
       description: 'Shows the amount of warnings a member has.',
@@ -27,14 +30,14 @@ export default class WarningsCommand extends Command {
 
     const userID = ctx.args.get(0);
     const u = findUser(this.bot, userID);
-    if (!u || u === undefined) return ctx.sendTranslate('global.unableToFind');
+    if (!u) return ctx.sendTranslate('global.unableToFind');
 
     const member = ctx.guild!.members.get(u.id);
     if (!member) return ctx.sendTranslate('commands.moderation.notInGuild', {
       user: `${u.username}#${u.discriminator}`
     });
 
-    const settings = await this.bot.warnings.get(ctx.guild!.id, member.id);
+    const settings = await this.warningService.get(ctx.guild!.id, member.id);
 
     return !settings
       ? ctx.sendTranslate('commands.moderation.warnings.none')
