@@ -1,33 +1,42 @@
-import { injectable, inject } from 'inversify';
-import { Guild, TextChannel } from 'eris';
+import { inject, injectable } from 'inversify';
+import { Client, Guild, TextChannel } from 'eris';
 import { TYPES } from '../types';
-import Client from '../structures/Bot';
+import Bot from '../structures/Bot';
 import Event from '../structures/Event';
 import { createEmptyEmbed } from '../util/EmbedUtils';
+import GuildSettingsService from '../structures/services/settings/GuildSettingsService';
+import StatusManager from '../structures/managers/StatusManager';
+import CommandStatisticsManager from '../structures/managers/CommandStatisticsManager';
 
 @injectable()
 export default class GuildJoinedEvent extends Event {
-  constructor(@inject(TYPES.Bot) client: Client) {
-    super(client, 'guildCreate');
+  constructor(
+      @inject(TYPES.Bot) bot: Bot,
+      @inject(TYPES.Client) private client: Client,
+      @inject(TYPES.GuildSettingsService) private guildSettingsService: GuildSettingsService,
+      @inject(TYPES.StatusManager) private statusManager: StatusManager,
+      @inject(TYPES.CommandStatisticsManager) private commandStatisticsManager: CommandStatisticsManager
+  ) {
+    super(bot, 'guildCreate');
   }
 
   async emit(guild: Guild) {
-    this.bot.settings.create(guild.id);
+    await this.guildSettingsService.create(guild.id);
     this.bot.logger.info(`Joined guild ${guild.name} (${guild.id})`);
 
-    this.bot.status.updateStatus();
-    this.bot.statistics.guildCount++;
-    await this.bot.redis.set('guilds', this.bot.client.guilds.size);
+    this.statusManager.updateStatus();
+    this.commandStatisticsManager.guildCount++;
+    await this.bot.redis.set('guilds', this.client.guilds.size);
 
-    const channel = await this.bot.client.getRESTChannel('529593466729267200');
+    const channel = this.client.getChannel('529593466729267200');
     if (channel && channel.type === 0) {
       const chan = (channel as TextChannel);
       const embed = createEmptyEmbed()
         .setAuthor(`| Joined ${guild.name} (${guild.id})`, undefined, this.bot.client.user.dynamicAvatarURL('png', 1024))
-        .setFooter(`Now at ${this.bot.client.guilds.size} Guilds`)
+        .setFooter(`Now at ${this.client.guilds.size} Guilds`)
         .build();
 
-      chan.createMessage({ embed });
+      await chan.createMessage({ embed });
     }
   }
 }
