@@ -7,13 +7,17 @@ import findUser from '../../util/UserUtil';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
 import Bot from '../../structures/Bot';
+import WarningService from '../../structures/services/WarningService';
+import PunishmentService from '../../structures/services/PunishmentService';
 
 @injectable()
 export default class PardonCommand extends Command {
   constructor(
-    @inject(TYPES.Bot) client: Bot
+    @inject(TYPES.Bot) bot: Bot,
+    @inject(TYPES.WarningService) private warningService: WarningService,
+    @inject(TYPES.PunishmentService) private punishmentService: PunishmentService
   ) {
-    super(client, {
+    super(bot, {
       name: 'pardon',
       description: 'Pardon a member from this guild',
       usage: '<user> <amount>',
@@ -31,7 +35,7 @@ export default class PardonCommand extends Command {
 
     const userID = ctx.args.get(0);
     const user = findUser(this.bot, userID);
-    if (!user || user === undefined) return ctx.sendTranslate('global.unableToFind');
+    if (!user) return ctx.sendTranslate('global.unableToFind');
 
     const member = ctx.guild!.members.get(user.id);
     const amount = Number(ctx.args.get(1));
@@ -40,9 +44,10 @@ export default class PardonCommand extends Command {
     });
 
     if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.sendTranslate('global.hierarchy');
+    else if (!PermissionUtils.above(ctx.me, member)) return ctx.sendTranslate('global.botHierarchy');
 
-    await this.bot.punishments.pardon(member!, amount);
-    const warns = await this.bot.warnings.get(ctx.guild!.id, member.id);
+    await this.punishmentService.pardon(member!, amount);
+    const warns = await this.warningService.get(ctx.guild!.id, member.id);
 
     return warns === null 
       ? ctx.sendTranslate('commands.moderation.pardon.noWarnings', { user: `${user.username}#${user.discriminator}` })
