@@ -1,5 +1,5 @@
-import { injectable, inject } from 'inversify';
-import { User, Member } from 'eris';
+import { inject, injectable } from 'inversify';
+import { Member, User } from 'eris';
 import { TYPES } from '../types';
 import Event from '../structures/Event';
 import Bot from '../structures/Bot';
@@ -7,7 +7,7 @@ import Bot from '../structures/Bot';
 @injectable()
 export default class UserUpdateEvent extends Event {
   constructor(
-    @inject(TYPES.Bot) client: Bot
+      @inject(TYPES.Bot) client: Bot
   ) {
     super(client, 'userUpdate');
   }
@@ -22,6 +22,6 @@ export default class UserUpdateEvent extends Event {
   }
 
   async emit(user: User) {
-    for (const member of this.getMutualGuilds(user)) this.bot.automod.handleMemberNameUpdate(member);
+    for (const member of this.getMutualGuilds(user)) await this.bot.automod.handleMemberNameUpdate(member);
   }
 }
