@@ -44,7 +44,7 @@ export default class ArgumentParser {
    * @param start The start of the slicing process
    * @param end The end process
    */
-  slice(start: number, end?: number) {
+  slice(start?: number, end?: number) {
     return this.args.slice(start, end);
   }
 }
