@@ -1,6 +1,8 @@
 import model, { WarningModel } from '../../models/WarningSchema';
 import { WarningBase as Base } from './settings/SettingsBase';
+import { injectable } from 'inversify';
 
+@injectable()
 export default class WarningService implements Base<WarningModel> {
   public model = model;
 
@@ -8,18 +10,26 @@ export default class WarningService implements Base<WarningModel> {
     const document = await this.model
       .findOne({ guild: guildId, user: userId })
       .exec();
-    if (!document || document === null) return null;
+    if (!document) return null;
     return document;
   }
 
-  create(guildId: string, userId: string) {
+  async create(guildId: string, userId: string) {
     const query = new this.model({
       guild: guildId,
       user: userId,
-      amount: 1,
+      amount: 0,
     });
-    query.save();
+    await query.save();
     return query;
+  }
+
+  async getOrCreate(guildId: string, userId: string) {
+    const warnings = await this.get(guildId, userId);
+    if (!warnings) {
+      return this.create(guildId, userId);
+    }
+    return warnings;
   }
 
   remove(guildId: string, userId: string) {
